+++ packages/artinchip/lvgl-ui/lvgl_v9/lvgl/src/others/sysmon/lv_sysmon.c
@@ -274,14 +274,14 @@ static void mem_observer_cb(lv_observer_t * observer, lv_subject_t * subject)
     lv_obj_t * label = lv_observer_get_target(observer);
     const lv_mem_monitor_t * mon = lv_subject_get_pointer(subject);
 
-    size_t used_size = mon->total_size - mon->free_size;;
-    size_t used_kb = used_size / 1024;
-    size_t used_kb_tenth = (used_size - (used_kb * 1024)) / 102;
-    size_t max_used_kb = mon->max_used / 1024;
-    size_t max_used_kb_tenth = (mon->max_used - (max_used_kb * 1024)) / 102;
+    unsigned long used_size = mon->total_size - mon->free_size;;
+    unsigned long used_kb = used_size / 1024;
+    unsigned long used_kb_tenth = (used_size - (used_kb * 1024)) / 102;
+    unsigned long max_used_kb = mon->max_used / 1024;
+    unsigned long max_used_kb_tenth = (mon->max_used - (max_used_kb * 1024)) / 102;
     lv_label_set_text_fmt(label,
-                          "%zu.%zu kB (%d%%)\n"
-                          "%zu.%zu kB max, %d%% frag.",
+                          "%lu.%lu kB (%d%%)\n"
+                          "%lu.%lu kB max, %d%% frag.",
                           used_kb, used_kb_tenth, mon->used_pct,
                           max_used_kb, max_used_kb_tenth,
                           mon->frag_pct);
